Fix constants in signal.h
